requirements.txt
Flask~=3.1.1
Werkzeug~=3.1.3
Flask-Migrate~=4.1.0
Flask-JWT-Extended~=4.7.1
flask-cors~=5.0.1
SQLAlchemy~=2.0.41
alembic~=1.15.2
PyJWT~=2.10.1
pymysql~=1.1.0
bcrypt~=4.3.0
marshmallow~=4.0.0
pydantic~=1.10.22
python-dotenv~=1.1.0
